Use storage path if disks don't have valid serial numbers
Some disk devices (virtio) don't have serial numbers by default.
This breaks curtin which requires a serial or a path to uniquely
find the target disks.  Add a fallback to device path if the
current device includes the eye catcher 'Unknown Serial' which
is set by probert for devices that do not have a serial number.

Signed-off-by: Ryan Harper <ryan.harper@canonical.com>