Don't require root for disk info
The spec calls for invoking hdparm -i on the device, but not all of our
devices will display hdparm data (virtual, ceph, or other such devices).
Additionally the bulk if the data in hdparm itself is not really useful
in helping users identify a specific device.  Instead, display a short
set of helpful info about the device:
  - /dev/XXXX
  - Vendor
  - Model
  - Serial
  - Size (in human-readable and bytes)
  - devpath (/devices/XXXX)
  - rotational (true or false)

The current UX requires us to select another disk after returning to
the list, with this patch though we enable switching between the
available disks via 'n' and 'p', or 'j' and 'k', which will be switched
to 'tab' and 'shift tab'.

Signed-off-by: Ryan Harper <ryan.harper@canonical.com>