Implement username field protection
man 8 useradd says:

    It is usually recommended to only use usernames that begin with a
    lower case letter or an underscore, followed by lower case
    letters, digits, underscores, or dashes. They can end with a
    dollar sign. In regular expression terms: [a-z_][a-z0-9_-]*[$]?

    On Debian, the only constraints are that usernames must neither
    start with a dash ('-') nor plus ('+') nor tilde ('~') nor
    contain a colon (':'), a comma (','), or a whitespace (space: ' ',
    end of line: '\n', tabulation: '\t', etc.).  Note that using a
    slash ('/') may break the default algorithm for the definition
    of the user's home directory.

    Usernames may only be up to 32 characters long.

In this patch we implement most of this.  Subset of the regular
expression suggested above is used to limit input into the username
field.  We've not yet determined how to provide a fixed width widget
so at this time, user can input more than 32 characters, but upon
selectin done, we raise and error and reset the state.

Signed-off-by: Ryan Harper <ryan.harper@canonical.com>