Overhaul networkin model and networking actions
The network model wasn't not prepared to allow manual configuration
to affect the state such that after manual configuration the view
would be able to extract the current state of a network device
config.  Reuse the Blockdev model as Networkdev as well as encapsulating
the use of the action objects as a means of storting the current
configuration.

Adjust the views to use the updated model.

Signed-off-by: Ryan Harper <ryan.harper@canonical.com>