Add initial bonding UX panels and model connections
We now have a bonding UX panel, modeled after the raid menu
where we have a list of interfaces and a radiobox for
selecting the bond mode.  This calls into the network
model to add the bond interface.

- Update the network view to only display the Bond option if we have
more than one interface that is available (ie, not already part of
a bond).
- Also update the bond info to include the bond interface composition
in the hardware info such that it displays according to spec
- Add iface_is_bonded_slave to model
- Update network model info dict to include bond_{slave,master}.  This
helps us display the (Bonded) attribute for slaves, but not the primary
bond iface.

Signed-off-by: Ryan Harper <ryan.harper@canonical.com>