Fix curtin install error detection and presentation
When curtin encountered an error, if the resulting message
was very large, this would prevent the progress screen from
displaying useful information and left the user believing the
install was still going on even when it had stopped.

Resolve this by applying some post processing on the curtin
error message to help user better spot the issue.  We also
update the Title, Header and Footer to indicate an error has
occurred.

Finally, left-align the progress output.

Signed-off-by: Ryan Harper <ryan.harper@canonical.com>