amber is now dark cyan
So on VT background brown looks bad, and background yellow is not
supported... until after a vtswitch. Not sure how to cause a clean
vtswitch of the framebuffer and instead use dark cyan, which is a nice
"cold" color. This is inline with probably future rename of cancel
buttons into back buttons.