store curtin's output in journald
This changes to put curtin's output into journald and use a
systemd.journal.Reader instance to read it. It's a bit simpler than the current
code, which is nice. It also allows us to play with displaying kernel or other
messages intermixed with the curtin output, which might be interesting to
explore.

This does mean that the curtin output is not on disk anywhere, but as it's
pretty trivially recoverable with `journalctl -t curtin_output -o cat` I think
that's OK.