remove the game of running curtin twice in favour of new curtin features
subiquity previously ran curtin twice, once to do the install and the second
time to configure the user data. All this second run did was to drop some
files into the fs and now that we can tell curtin not to unmount everything
when it is done, this can be achieved much more simply. So do that, and rip out
a bunch of the complexity having two curtin configs etc brings.