remove curtin from the snap
livecd-rootfs puts it in the install media for us. this could go the other way
and remove it from the install media but eh, i don't much care and having two
copies is silly.

this requires removing SNAP from the environment before invoking curtin,
because curtin uses SNAP to find bits of itself (this probably means that we
were using a mix of the two copies of curtin we had before this change!)