Implement zfsroot support
Curtin supports a ZFS on Root mode where users can configure their rootfs
as a ZFS filesystem.  This is experimental and restricts use of ZFS only
to the rootfs volume.  This patch adds 'zfsroot' to the supported filesystem
list and enforces the requirements of only one zfsroot and only mounted at
'/'.  Add some code to blank out zfsroot if one is already selected.

This patch also adds a guided answer file for zfsroot.  One small change
to the guided path allowing the answers to supply the filesystem value for
the rootfs.