Fix curtin block comments in example events.
I fixed the 'c u r t i n  b l o c k'... report event in the code a few
commits back, but when using dry-run it was still present because
dry run uses the events json rather than real curtin output.

This just updates the events json to be in line with reality as
it was confusing to see it in the dry run path.