services: run console-conf after core18.start-snapd.service
On core18 the firstboot seeding happens via the core18.start-snapd
service. This service will do the initial start of snapd when the
system is unseeded. This service will wait until the system is
fully seeded.

We cannot run console-conf before that because it uses the "snap"
command which will not be available before the system is seeded.

Once that has landed https://github.com/snapcore/core18/pull/74
can be reverted (in fact the hooks/200-console-conf-after.chroot
file can be removed entirely).