console-conf-wrapper: deal more gracefully with seeding failures
Right now the console-conf wrapper assumes that the "snap" command
is available. However in a core18 world this may not be the case.

The snap command is provided by the snapd snap which needs to get
seeded first. If this seeding fails no snap command is available.

This PR changes console-conf-wrapper to deal more gracefully with
the error condition by at least showing the relevant log and
explaining what is going on to the user.