Use attribute metadata to drive maintenance of backlinks
By backlinks I mean for example the link a filesystem has to a mount
that uses it, or the link from a partition to the RAID it is a member
of.

This might be a little magical, but something like this is really nice
when interpreting config objects from curtin.