show progress when apply_autoinstall_config takes >0.1s
If, when the overall installer is interactive but a particular
controller is not and applying the config for that installer takes more
than 0.1 seconds, show the progress view until it completes (but ensure
that the progress view is on screen for at least a whole second if it is
shown at all).

Block all input during that (up to) 0.1s wait.

Send events from non-interacttive controllers to the progress view.