add and use the concept of a "global overlay"
a problem with my previous merge was that the top overlay was always
removed when the install finished, but if the user had opened a help
topic in the mean time that was the wrong thing to do. also if the
view was switched under the install notification (which is a bit
unlikely but could happen) the dialog would disappear.

this finally pushed me to generically handle "global overlays" that are
not tied to the underlying, controller-controlled view. there are a few
of these now:

 * help dialogs
 * error reports
 * the "are you sure?" install confirmation
 * now the install notification

this adds an api for adding and removing them, and the api for removing
them includes which overlay to remove.