refresh: default to updating the installer when available
Way too many good bugfixes are shipped in the new installer, thus
default to update the installer when a new one is available.

Also, it seems like self.active was always interactive, meaning
autoinstall refresh-installer doesn't do anything? Not sure if that
was intentional. Because self.active was only assigned in __init__,
which always has self.ai_data as an empty dict.