console-conf: show journal log during install
Display the snapd journal while the system installs. This will
give us important clues what might go wrong during a install
and also gives the user some sense of progress.

For post-beta we need to think if we really want to display this
level of detail. OTOH we don't expect people to watch the console
during firstboot yet and if that will happen we need to also tweak
a bunch more (like kernel messages, systemd messages etc).