snapcraft: unminimize iso-codes package
When building the snap in a clean environment, we start with a minimised core image. This image excludes translation
files, which are needed for the "languagelists" step. The following error will show up when building:

    Building languagelists
    do not know native name for oc
    Failed to run 'override-build': Exit code was 1.

This is due to translation files being excluded by default in /etc/dpkg/dpkg.cfg.d/excludes on minimised systems:

    # Drop all translations
    path-exclude=/usr/share/locale/*/LC_MESSAGES/*.mo

We now comment this line and reinstall the "iso-codes" package.