Makefile: add xorriso and isolinux to install_deps
In order to use the inject-subiquity-snap.sh script, we need some extra dependencies: xorriso and isolinux.

When xorriso is missing, the following error occurs:

    ./scripts/inject-subiquity-snap.sh: line 184: xorriso: command not found

When isolinux is missing, the image build step for x86-64 is failing:

    xorriso : FAILURE : Given path does not exist on disk: -boot_image system_area='/usr/lib/ISOLINUX/isohdpfx.bin'

This file is not present in the source ISO image, so it has to come from the isolinux package.