catch all errors getting application status
When refreshing the snap, it is possible for the client to make a "get
current status" request to the new server before it gets restarted and
the new server can return values the client does not know about. Rather
than crashing the client when this happens, just ignore all errors when
getting application status.