Integration fixes, schema update.
Update schema for valid timezones.
Add client_variant API - client must call this exactly once to help set
the right expectations on the controllers to wait on (or not).
Remove underscore from geoip interface properties.
Update geoip interface to be non-blocking - if we ask too early, it just
returns default answers.  This should be a non-common case due to the
timing of when the request happens, and how quickly it should occur.
Fix subprocess usage for python 3.6.