Add client_variant (#976)
We need to know which controllers to block on, and which are optional.
The mechanism, for now, is this client_variant functionality, which adds
a mandatory early API call that the client shall use.  This allows
controllers to say that they are only required to be configured for
certain install clients, such as is necessary for timezone (which is
needed on desktop, and optional on server).