Abstract resource file lookup relative to SNAP (#998)
* Abstract resource file lookup relative to SNAP

Using SNAP to find the resource files is a bit of an odd requirement for
non-subiquity clients.  Start abstracting that away so it's easier to
change later.

Also move loadkeys to lookup from this, since the previous solution is
still a problem for ubuntu-desktop-installer.

* Move loadkeys / configure-apt to bin

loadkeys / configure-apt are in usr/bin only when we specially put them
there by way of snapcraft, which makes things a little harder on
non-subiquity clients.  Move them to bin, which is to say don't have
snapcraft put them in usr/bin.