Use inheritance for IdentityForm
The previous composition version had some flaws due to multiple
instances of a given Field with different addresses when reinstalled in
the parent class. This caused the field validation validate_ to not work
as it was not taken the expected referenced values.

Instead of duplicating the code, we use an inherited version, taking and
overriding only our desired fields.

We still need to derive WSLIdentityView from the base class instead of
IdentityView, as this is what is instantiating the form.
(WSLIdentityForm instead of IdentityForm).

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>