storage/v2: reuse existing partitions
Add test for flow of reusing existing partitions.  Certain operations
are allowed, such as formatting and using an existing partition, or
mounting a partition without formatting it.  Creating new in free space
is off limits, as is deleting a partition (even if it is just about to
be added again).