models/subiquity: add support for merging lists when merging user-data
The current merge_config() function from curtin does not merge lists in any way, but overwrites them. This is generally
a good idea, but in case of user-data we may want to merge lists instead.

One of the possible use cases is defining the main user in the "identity" section, and some extra users in the
user-data "users" section.

Before, the user defined in the "identity" section overwrote any users defined in user-data. With this change, the user
list can now be automatically merged, with user-data having lower priority than autoinstall configuration.

An example of such a configuration:

    #cloud-config
    autoinstall:
      version: 1
      ...
      identity:
        hostname: testname
        realname: Real Name
        username: someuser
        password: "$6$..."
      user-data:
        users:
          - name: extrauser
            gecos: Extra User
            ssh_authorized_keys:
              - "ssh-ed25519 AAAA..."
            shell: /bin/bash