test/api: make less flaky
This is part of a changeset intended to have the API tests run in
parallel.  I don't consider this ideal - I still see failures when
running tests in parallel - but am providing it now with the intent of
helping people who are seeing api test flakiness even when those tests
are run serially.

* check for socket existence on startup
* lengthen timeouts on shutdown