Compare the snaps versions to avoid suggesting a downgrade
In specific scenarios, the version of the snap as returned by snapd is
older than the version currently running. When that happens, we were
suggesting the user to update. This action would effectively result in a
downgrade.

We now attempt to determine if the snapd version is greater than the one
currently running. If we can't the behavior is unaffected.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>